Changed layout of user profile to include rosary, journal entry and feedback submitted count, also shows last 5 feedbacks user submits.
Changed saint info that gets fetched in saint screen to not include every field and only the neccessary for the page. Full saint info is fetched in saint detal modal when opened.

Clarified chapters in ToS and Privacy policy.

Optimised screens to be able to handle paginated fetches from backend to not load everything at once.